fix: split profile header
