+++ frontend/src/pages/profile/ProfileHeader.tsx
@@ -1,0 +1,70 @@
+import { Button } from '@material-tailwind/react';
+import { HomeIcon, MapPinIcon } from '@heroicons/react/24/solid';
+import type { Personnel } from '../dashboard';
+
+function HorizontalLine() {
+  return (
+    <svg
+      xmlns="http://www.w3.org/2000/svg"
+      fill="none"
+      strokeWidth={1}
+      stroke="gray"
+      className="h-10 w-10"
+    >
+      <path id="triangle" d="M1 5 L0 35" />
+    </svg>
+  );
+}
+
+const ProfileHeader = ({ personnel }: { personnel: Personnel }) => {
+  return (
+    <div className="flex items-center flex-row pb-2 pr-12">
+      <div className="flex items-center pr-12">
+        <h2 className="font-semibold">
+          {personnel.firstName} {personnel.lastName}
+        </h2>
+        <div className="pl-4">
+          <span className="bg-activeGreen px-2 rounded-full">
+            {personnel.active === true ? 'Active' : 'Inactive'}
+          </span>
+        </div>
+      </div>
+      <div>
+        <HorizontalLine />
+      </div>
+      <div className="pl-16">
+        <div className="flex flex-row">
+          <MapPinIcon className="h-7 w-7" />
+          <div className="pl-2">
+            <p className="text-textGray">Work Location</p>
+            <p>
+              {personnel.workLocation}, {personnel.region}
+            </p>
+          </div>
+        </div>
+      </div>
+      <div className="pl-12 grow">
+        <div className="flex flex-row">
+          <HomeIcon className="h-7 w-7" />
+          <div className="pl-2">
+            <p className="text-textGray">Home Location</p>
+            <p>
+              {personnel.workLocation}
+              {/* TODO: When Home Location is made on backend, use this */}
+            </p>
+          </div>
+        </div>
+      </div>
+      <div>
+        <Button
+          placeholder="Edit Info"
+          className="bg-blue font-thin capitalize text-base"
+        >
+          Edit Info
+        </Button>
+      </div>
+    </div>
+  );
+};
+
+export default ProfileHeader;
